fix(database/migration/create_books_table): update down method

+++ database/migrations/1734615758072_create_books_table.ts
@@ -21,6 +21,9 @@ export default class extends BaseSchema {
   }
 
   async down() {
-    this.schema.dropTable(this.tableName)
+    const hasTable = await this.db.schema.hasTable(this.tableName)
+    if (hasTable) {
+      this.schema.dropTable(this.tableName)
+    }
   }
 }
